+++ js/jquery.scrollmagic.debug.js
@@ -1,5 +1,5 @@
 /*
-ScrollMagic v1.1.2
+ScrollMagic v1.2.0
 The jQuery plugin for doing magical scroll interactions.
 (c) 2014 Jan Paepke (@janpaepke)
 License & Info: http://janpaepke.github.io/ScrollMagic
@@ -12,11 +12,19 @@ Greensock License info at http://www.greensock.com/licensing/
 */
 /*
 	@overview Debug Extension for ScrollMagic.
-	@version	1.1.2
+	@version	1.2.0
 	@license	Dual licensed under MIT license and GPL.
 	@author		Jan Paepke - e-mail@janpaepke.de
 */
-(function($, ScrollScene) {
+(function (root, factory) {
+    if (typeof define === 'function' && define.amd) {
+        // AMD. Register as an anonymous module.
+        define(['ScrollScene', 'jquery'], factory);
+    } else {
+    		// no browser global needed, just execute
+        factory(root.ScrollScene, root.jQuery);
+    }
+}(this, function(ScrollScene, $) {
 	/**
 	 * Add Indicators for a ScrollScene.  
 	 * __REQUIRES__ ScrollMagic Debug Extension: `jquery.scrollmagic.debug.js`  
@@ -248,4 +256,4 @@ Greensock License info at http://www.greensock.com/licensing/
 			}
 		}
 	};
-})(jQuery, ScrollScene);
+}));